requirements.txt
APScheduler==3.10.4
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
cssselect2==0.7.0
Django==4.2.7
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-filter==23.4
django-jsonfield==1.4.1
django-mathfilters==1.0.0
django-model-utils==4.3.1
django-simple-history==3.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
freetype-py==2.3.0
html5lib==1.1
idna==3.6
Jinja2==3.1.2
jsonfield==3.1.0
lxml==4.9.3
MarkupSafe==2.1.3
numpy==1.26.2
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.1.3
Pillow==10.1.0
pycairo==1.25.1
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.2
PyJWT==2.8.0
pypdf==3.17.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.7
requests==2.31.0
rlPyCairo==0.3.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
swapper==1.3.0
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
utils==1.0.1
webencodings==0.5.1
xhtml2pdf==0.2.13
XlsxWriter==3.1.9
